Add `ignoreSlackText` + `ignoreSlackTitle`
